move Test 用に作成したサンプル構造体を移動しました.
